[REFACTOR] 마이페이지 대시보드 데이터 get api 리팩터링

+++ src/main/java/com/vivid/apiserver/domain/individual_video/application/IndividualVideoService.java
@@ -74,7 +74,7 @@ public class IndividualVideoService {
     /**
      * individual video의 진행도를 업데이트하는 메소드
      */
-    public void updateProgressRate(String individualVideoId, Long progressRate) {
+    public void updateProgressRate(String individualVideoId, Integer progressRate) {
 
         IndividualVideo individualVideo = individualVideoQueryService.findById(individualVideoId);
